requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncssh==2.14.2
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pycparser==2.22
pyTelegramBotAPI==4.17.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
